Renamed cluster_metadata into cluster_info.
